Changed the type of the file and rank in position to int from byte

+++ internal/move/move.go
@@ -5,5 +5,5 @@ type Move struct {
 }
 
 type Position struct {
-	File, Rank byte
+	File, Rank int
 }
